+++ avian3d/src/lib.rs
@@ -26,8 +26,8 @@ use bevy_tnua_physics_integration_layer::data_for_backends::{
 };
 use bevy_tnua_physics_integration_layer::obstacle_radar::TnuaObstacleRadar;
 use bevy_tnua_physics_integration_layer::subservient_sensors::TnuaSubservientSensor;
-use bevy_tnua_physics_integration_layer::TnuaPipelineStages;
-use bevy_tnua_physics_integration_layer::TnuaSystemSet;
+use bevy_tnua_physics_integration_layer::TnuaPipelineSystems;
+use bevy_tnua_physics_integration_layer::TnuaSystems;
 
 /// Add this plugin to use avian3d as a physics backend.
 ///
@@ -62,7 +62,7 @@ impl Plugin for TnuaAvian3dPlugin {
     fn build(&self, app: &mut App) {
         app.configure_sets(
             self.schedule,
-            TnuaSystemSet
+            TnuaSystems
                 // Need to run _before_ `First`, not after it. The documentation is misleading. See
                 // https://github.com/Jondolf/avian/issues/675
                 .before(PhysicsStepSystems::First)
@@ -75,11 +75,11 @@ impl Plugin for TnuaAvian3dPlugin {
                 update_proximity_sensors_system,
                 update_obstacle_radars_system,
             )
-                .in_set(TnuaPipelineStages::Sensors),
+                .in_set(TnuaPipelineSystems::Sensors),
         );
         app.add_systems(
             self.schedule,
-            apply_motors_system.in_set(TnuaPipelineStages::Motors),
+            apply_motors_system.in_set(TnuaPipelineSystems::Motors),
         );
         app.add_systems(
             Update,