+++ demos/src/bin/shooter_like.rs
@@ -43,7 +43,7 @@ use tnua_demos_crate::ui::component_alterbation::CommandAlteringSelectors;
 use tnua_demos_crate::ui::info::InfoSource;
 #[cfg(feature = "egui")]
 use tnua_demos_crate::ui::plotting::PlotSource;
-use tnua_demos_crate::ui::DemoInfoUpdateSystemSet;
+use tnua_demos_crate::ui::DemoInfoUpdateSystems;
 use tnua_demos_crate::util::animating::{animation_patcher_system, GltfSceneHandler};
 
 fn main() {
@@ -104,7 +104,7 @@ fn main() {
     #[cfg(feature = "egui")]
     app.add_systems(
         Update,
-        character_control_info_dumping_system.in_set(DemoInfoUpdateSystemSet),
+        character_control_info_dumping_system.in_set(DemoInfoUpdateSystems),
     );
     app.add_systems(Update, character_control_radar_visualization_system);
     app.add_plugins(tnua_demos_crate::ui::DemoUi::<
@@ -131,7 +131,7 @@ fn main() {
             ScheduleToUse::Update => Update.intern(),
             ScheduleToUse::FixedUpdate => FixedUpdate.intern(),
         },
-        apply_platformer_controls.in_set(TnuaUserControlsSystemSet),
+        apply_platformer_controls.in_set(TnuaUserControlsSystems),
     );
     app.add_systems(Update, animation_patcher_system);
     app.add_systems(Update, animate_platformer_character);