+++ physics-integration-layer/CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## 0.10.0 - 2025-10-14
+### Changed
+- Upgrade to Bevy 0.17.
+- Rename `TnuaSystemSet` to `TnuaSystems`.
+- Rename `TnuaPipelineStages` to `TnuaPipelineSystems`.
+
 ## 0.9.0 - 2025-09-26
 ### Added
 - `cast_shape_rotation` field for `TnuaProximitySensor`.